6 - Remove setParameter(... Type type) methods in favour of setParameter( AllowableParameterType type)
