+++ hibernate-core/src/main/java/org/hibernate/persister/collection/NamedQueryCollectionInitializer.java
@@ -14,6 +14,7 @@ import org.hibernate.engine.spi.SharedSessionContractImplementor;
 import org.hibernate.internal.CoreLogging;
 import org.hibernate.internal.CoreMessageLogger;
 import org.hibernate.loader.collection.CollectionInitializer;
+import org.hibernate.metamodel.model.domain.AllowableParameterType;
 import org.hibernate.query.sql.spi.NativeQueryImplementor;
 
 /**
@@ -42,11 +43,11 @@ public final class NamedQueryCollectionInitializer implements CollectionInitiali
 			nativeQuery.setParameter(
 					nativeQuery.getParameterMetadata().getNamedParameterNames().iterator().next(),
 					key,
-					persister.getKeyType()
+					(AllowableParameterType) persister.getKeyType()
 			);
 		}
 		else {
-			nativeQuery.setParameter( 1, key, persister.getKeyType() );
+			nativeQuery.setParameter( 1, key, (AllowableParameterType) persister.getKeyType() );
 		}
 
 		nativeQuery.setCollectionKey( key ).setHibernateFlushMode( FlushMode.MANUAL ).list();
