+++ hibernate-core/src/test/java/org/hibernate/test/sql/hand/query/NativeSQLQueriesTest.java
@@ -149,28 +149,6 @@ public class NativeSQLQueriesTest extends BaseCoreFunctionalTestCase {
 			s.close();
 		}
 	}
-	
-	@Test
-	public void testRegisteredNamedSQLQueryWithScalar()
-	{
-		final NamedNativeQueryMementoBuilder builder = new NamedNativeQueryMementoBuilder();
-		builder.setName("namedQuery");
-		builder.setQuery("select count(*) AS c from ORGANIZATION");
-		builder.setQueryReturns(new NativeSQLQueryReturn[1]);
-		
-		sessionFactory().registerNamedSQLQueryDefinition("namedQuery", builder.createNamedQueryDefinition());
-
-		final Session s = openSession();
-		s.beginTransaction();
-		final NativeQuery query = (NativeQuery) s.getNamedQuery("namedQuery");
-		query.addScalar("c");
-		final Number result = (Number) query.uniqueResult();
- 		s.getTransaction().commit();
-		s.close();
-		
-		assertNotNull(result);
-		assertTrue(0 == result.intValue());
-	}
 
 	@Test
 	public void testManualSynchronization() {
